+++ parse/file.go
@@ -16,18 +16,6 @@ func fileGoType(x ast.Expr) string {
 	switch t := x.(type) {
 	case *ast.StarExpr:
 		return "*" + fileGoType(t.X)
-	case *ast.ArrayType:
-		switch l := t.Len.(type) {
-		case nil:
-			return "[]" + fileGoType(t.Elt)
-		case *ast.BasicLit:
-			return fmt.Sprintf("[%s]", l.Value) + fileGoType(t.Elt)
-		default:
-			panic(fmt.Errorf("reform: fileGoType: unhandled array length %#v (%#v). Please report this bug.", l, t))
-		}
-		// return "[]" + fileGoType(t.Elt)
-	case *ast.SelectorExpr:
-		return fileGoType(t.X) + "." + t.Sel.String()
 	case *ast.Ident:
 		return t.String()
 	default:
@@ -75,22 +63,21 @@ func parseStructTypeSpec(ts *ast.TypeSpec, str *ast.StructType) (*StructInfo, er
 		if column == "" {
 			return nil, fmt.Errorf(`reform: %s has field %s with invalid "reform:" tag value, it is not allowed`, res.Type, name.Name)
 		}
-		typ := fileGoType(f.Type)
-		if isPK && strings.HasPrefix(typ, "*") {
-			return nil, fmt.Errorf(`reform: %s has pointer field %s with with "pk" label in "reform:" tag, it is not allowed`, res.Type, name.Name)
+		var pkType string
+		if isPK {
+			pkType = fileGoType(f.Type)
+			if strings.HasPrefix(pkType, "*") {
+				return nil, fmt.Errorf(`reform: %s has pointer field %s with with "pk" label in "reform:" tag, it is not allowed`, res.Type, name.Name)
+			}
+			if res.PKFieldIndex >= 0 {
+				return nil, fmt.Errorf(`reform: %s has field %s with with duplicate "pk" label in "reform:" tag (first used by %s), it is not allowed`, res.Type, name.Name, res.Fields[res.PKFieldIndex].Name)
+			}
 		}
-		if isPK && res.PKFieldIndex >= 0 {
-			return nil, fmt.Errorf(`reform: %s has field %s with with duplicate "pk" label in "reform:" tag (first used by %s), it is not allowed`, res.Type, name.Name, res.Fields[res.PKFieldIndex].Name)
-		}
-		// if isPKOrOmitEmpty && strings.HasPrefix(typ, "*") {
-		// 	return nil, fmt.Errorf(`reform: %s has pointer field %s with with "omitempty" label in "reform:" tag, it is not allowed`, res.Type, name.Name)
-		// }
 
 		res.Fields = append(res.Fields, FieldInfo{
 			Name:   name.Name,
-			Type:   typ,
+			PKType: pkType,
 			Column: column,
-			// PKOrOmitEmpty: isPKOrOmitEmpty,
 		})
 		if isPK {
 			res.PKFieldIndex = n